Major: Refactor streamdeck code to be less duplicated and have better error handling
