+++ lib/usb/elgato.js
@@ -16,138 +16,18 @@
  */
 
 
-var StreamDeck   = require('elgato-stream-deck').openStreamDeck;
 var util         = require('util');
 var debug        = require('debug')('lib/usb/elgato');
-var common       = require('./common');
+var elgato_base  = require('./elgato-base');
 
 function elgato(system, devicepath) {
 	var self = this;
 
-	self.info = {
-		type: 'Elgato Streamdeck device',
-		devicepath: devicepath,
-		device_type: 'StreamDeck',
-		config: [ 'brightness', 'orientation', 'page' ],
-		keysPerRow: 5,
-		keysTotal: 15
-	};
-
-	self.config = {
-		brightness: 100,
-		rotation: 0,
-		page: 1
-	};
-
-	process.on('uncaughtException', function (err) {
-		system.emit('log', 'device'+self.serialnumber+')', 'debug', 'Exception:' + err);
-	});
-	process.on('unhandledRejection', function (err) {
-		system.emit('log', 'device'+self.serialnumber+')', 'debug', 'Exception:' + err);
-	});
-
-	debug('Adding Elgato Streamdeck USB device', devicepath);
-
-	self.devicepath = devicepath;
-	self.streamDeck = new StreamDeck(devicepath, { useOriginalKeyOrder: true });
-	self.buttonState = [];
-
-	self.info.serialnumber = self.serialnumber = self.streamDeck.device.getDeviceInfo().serialNumber;
-
-	system.emit('log', 'device('+self.serialnumber+')', 'debug', 'Elgato Streamdeck detected');
-
-	// How many items we have left to load until we're ready to begin
-	self.loadingItems = 0;
-	self.system = system;
-
-	// send elgato ready message to devices :)
-	setImmediate(function() {
-		system.emit('elgato_ready', devicepath);
-	});
-
-	self.streamDeck.on('down', function (keyIndex) {
-		var key = self.reverseButton(keyIndex);
-
-		if (key === undefined) {
-			return;
-		}
-
-		self.buttonState[key].pressed = true;
-		self.system.emit('elgato_click', devicepath, key, true, self.buttonState);
-	});
-
-	self.streamDeck.on('up', function (keyIndex) {
-		var key = self.reverseButton(keyIndex);
-
-		if (key === undefined) {
-			return;
-		}
-
-		self.buttonState[key].pressed = false;
-		self.system.emit('elgato_click', devicepath, key, false, self.buttonState);
-	});
-
-	self.streamDeck.on('error', function (error) {
-		console.error(error);
-		system.emit('elgatodm_remove_device', devicepath);
-	});
-
-	// Initialize button state hash
-	for (var button = 0; button < global.MAX_BUTTONS; button++) {
-		self.buttonState[button] = {
-			pressed: false
-		};
-	}
-
-	common.apply(this, arguments);
-
-	self.clearDeck();
+	elgato_base.apply(this, [system, devicepath, 'Elgato Streamdeck']);
 
 	return self;
 }
 elgato.device_type = 'StreamDeck';
-
-elgato.prototype.setConfig = function (config) {
-	var self = this;
-	if (self.config.brightness != config.brightness && config.brightness !== undefined) {
-		self.streamDeck.setBrightness(config.brightness);
-	}
-
-	if (self.config.rotation != config.rotation && config.rotation !== undefined) {
-		self.config.rotation = config.rotation;
-		self.system.emit('device_redraw', self.devicepath);
-	}
-
-	if (self.config.page != config.page && config.page !== undefined) {
-		self.config.page = config.page;
-
-		// also handeled in usb.js
-		self.system.emit('device_redraw', self.devicepath);
-	}
-
-	self.config = config;
-};
-
-elgato.prototype.quit = function () {
-	var self = this;
-	var sd = self.streamDeck;
-
-	if (sd !== undefined) {
-		try {
-			this.clearDeck();
-		} catch (e) {}
-
-		// Find the actual streamdeck driver, to talk to the device directly
-		if (sd.device === undefined && sd.streamdeck !== undefined) {
-			sd = sd.streamdeck;
-		}
-
-		// If an actual streamdeck is connected, disconnect
-		if (sd.device !== undefined) {
-			sd.device.close();
-		}
-	}
-};
 
 elgato.prototype.draw = function(key, buffer, attempts) {
 	var self = this;
@@ -161,7 +41,7 @@ elgato.prototype.draw = function(key, buffer, attempts) {
 
 	attempts++;
 
-	var drawKey = self.mapButton(key);
+	var drawKey = self.toDeviceKey(key);
 
 	try {
 
@@ -185,45 +65,6 @@ elgato.prototype.draw = function(key, buffer, attempts) {
 	}
 }
 
-elgato.prototype.begin = function() {
-	var self = this;
-	self.log('elgato.prototype.begin()');
-
-	self.streamDeck.setBrightness(self.config.brightness);
-};
-
-elgato.prototype.mapButton = function(input) {
-	var self = this;
-	var map = "4 3 2 1 0 9 8 7 6 5 14 13 12 11 10".split(/ /);
-	var devkey = self.toDeviceKey(input);
-
-	if (devkey < 0) {
-		return -1;
-	}
-
-	return parseInt(map[devkey]);
-}
-
-elgato.prototype.reverseButton = function(input) {
-	var self = this;
-
-	var map = "4 3 2 1 0 9 8 7 6 5 14 13 12 11 10".split(/ /);
-	for (var pos = 0; pos < map.length; pos++) {
-		if (map[input] == pos) return self.toGlobalKey(pos);
-	}
-
-	return;
-};
-
-elgato.prototype.clearDeck = function() {
-	var self = this;
-	self.log('elgato.prototype.clearDeck()')
-
-	for (var x = 0; x < self.info.keysTotal; x++) {
-		self.streamDeck.clearKey(x);
-	}
-}
-
-util.inherits(elgato, common);
+util.inherits(elgato, elgato_base);
 
 exports = module.exports = elgato;
